ncov narrative: move redirect to separate file
