+++ redirects.js
@@ -1,4 +1,4 @@
-
+const helpers = require("./auspice/server/getDatasetHelpers");
 
 const setup = (app) => {
 
@@ -18,6 +18,29 @@ const setup = (app) => {
   aren't valid - redirect instead to the nCoV section of the splash page */
   app.route(["/narratives/ncov", "/narratives/ncov/sit-rep"])
     .get((req, res) => res.redirect('/#ncov'));
+
+  /*
+   * Redirect to translations of narratives if the client has
+   * set language preference and the translation is available
+   *
+   * Currently only used for nCoV narratives and requires the
+   * language code to be the second to last part of the url.
+   */
+  app.route("/narratives/ncov/sit-rep/*")
+    .get(async (req, res, next) => {
+      const acceptedLanguages = req.acceptsLanguages();
+      if (acceptedLanguages && !acceptedLanguages[0].startsWith("en")) {
+        const {source, prefixParts} = helpers.splitPrefixIntoParts(req.url);
+        prefixParts.splice(-1, 0, acceptedLanguages[0]);
+        const potentialNarrative = prefixParts.join("/");
+        const availableNarratives = await source.availableNarratives();
+        if (availableNarratives.includes(potentialNarrative)) {
+          res.redirect(302, "/narratives/" + potentialNarrative);
+        }
+
+      }
+      next('route');
+    });
 
 };
